Implement editor
The ascii image editor is a tool for creating and editing ascii images
(character matrices). It provides instruments for drawing and utilities
such as copying, importing, etc.
The ascii animation editor is an extenstion of the ascii image editor.
It provides tools for creation and manipulation of ascii animations
(arrays of ascii images).